packages/pilotplan/pyproject.toml
[project]
name = "pilotplan"
version = "0.2.0"
description = "Declarative planning engine for k3sp"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.0",
    "ruamel.yaml>=0.18.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

